+++ src/onegov.event/HISTORY.rst
@@ -7,6 +7,9 @@ Unreleased
 - Use hstore for tags.
   [msom]
 
+- Filter for current occurrences by default.
+  [msom]
+
 0.0.1 (2015-08-20)
 ~~~~~~~~~~~~~~~~~~~